requirements.txt
asgiref==3.4.1
Django==3.2.5
django-admin-sortable==2.2.4
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
asgiref==3.4.1
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.5
django-admin-sortable==2.2.4
django-import-export==2.5.0
django-pandas==0.6.4
django-phonenumber-field==5.2.0
et-xmlfile==1.1.0
kiwisolver==1.3.1
MarkupPy==1.14
numpy==1.21.1
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.1
phonenumberslite==8.12.28
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
